Fix blockchain tests.
