+++ ethcore/blockchain/Cargo.toml
@@ -27,7 +27,8 @@ rlp_derive = { path = "../../util/rlp-derive" }
 
 [dev-dependencies]
 env_logger = "0.5"
-ethcore = { path = "../", features = ["test-helpers"] }
 ethkey = { path = "../../accounts/ethkey" }
 keccak-hash = "0.1"
 rustc-hex = "1.0"
+tempdir = "0.3"
+kvdb-memorydb = "0.1"
